+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - After update to v0.1.0 the plugin need to be reconfigured!
 
+## [0.3.0] - (24.01.2026)
+
+## Changes
+
+- added ACL file to manually put in OpenWrt device
+- added Fan as a additional control accessory type for Radio and SSID
+- added possibility to change SSID name direct from Home app
+- config schema updated
+- readme updated
+- cleanup
+
 ## [0.2.0] - (23.01.2026)
 
 ## Changes